separate custum hook definition and context definition
now useTableCellFocus hook can be tested as a simple custom hook,
and useTableCellFocusContext is a context version of it.
