+++ webserver/src/useTableCellFocus.test.ts
@@ -1,6 +1,5 @@
 import {
   useTableCellFocus,
-  useTableCellFocusDispatcher,
 } from './useTableCellFocus';
 
 import { renderHook, act } from '@testing-library/react';
@@ -21,7 +20,6 @@ describe('useTableCellFocus Tests', () => {
 
   test('setFocus to same cell make focusMode "Editing"', () => {
     const { result } = renderHook(() => useTableCellFocus());
-    const { result: resultDispatcher } = renderHook(() => useTableCellFocusDispatcher());
     const args = {
       commonParentId: dummyCommonParentId,
       commonProcessType: dummyCommonProcessType,
@@ -32,8 +30,8 @@ describe('useTableCellFocus Tests', () => {
     };
     // NOTE
     // two ast() calls are needed to emulate two setFocus() calls
-    act(() => resultDispatcher.current.setFocus(args));
-    act(() => resultDispatcher.current.setFocus(args));
+    act(() => result.current.setFocus(args));
+    act(() => result.current.setFocus(args));
     expect(result.current.focusPosition?.commonParentId).toBe(dummyCommonParentId);
     expect(result.current.focusPosition?.commonProcessType).toBe(dummyCommonProcessType);
     expect(result.current.focusMode).toBe("Editing");
